frontend/requirements.txt
# Frontend Requirements for Journey Risk Management System

# Core Flask dependencies
Flask==2.2.3
Werkzeug==2.2.3
Flask-Assets==2.0
python-dotenv==1.0.0

# Dash components for interactive dashboards
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-leaflet==0.1.23
dash-core-components==2.0.0
dash-html-components==2.0.0
plotly==5.14.1

# Asset compilation
libsass==0.22.0
cssmin==0.2.0
# jsmin==3.0.1

# HTTP client for API communication
requests==2.28.2

# Data manipulation
pandas==1.5.3
numpy==1.24.2

# Additional utilities
polyline==2.0.0  # For decoding route polylines

# Date/time handling
python-dateutil==2.8.2

# JSON processing
simplejson==3.18.4
